feat: aff function decompress
